回收站功能

+++ public/js/common/L-blog.js
@@ -96,21 +96,14 @@ app.factory('categoriesAllService',['$http',function ($http) {
  */
 app.factory('documentAllService',['$http',function ($http) {
   return{
-    get:function (limit,page) {
-      return $http({
-        method:'POST',
-        url:'/admin/manage/document_manage/get_page_document',
-        data:$.param({
-          limit:limit,
-          page:page,
-        }),
-        headers:{'content-type':'application/x-www-form-urlencoded'}
-      });
-    },
+
+    /*
+    * 根据每页显示数、第几页来获取文档数据
+    * */
     postLimitAndPage:function (limit,page) {
       return $http({
         method:'POST',
-        url:'/admin/manage/document_manage/get_page_document',
+        url:'/admin/manage/document_manage/get_yes_display_document',
         data:$.param({
           limit:limit,
           page:page,
@@ -118,10 +111,27 @@ app.factory('documentAllService',['$http',function ($http) {
         headers:{'content-type':'application/x-www-form-urlencoded'}
       });
     },
+    /*
+    * 删除单篇文档
+    * */
     removeOneDocument:function (doc) {
       return $http({
         method:'POST',
         url:'/admin/manage/document_manage/remove_one_document',
+        data:$.param({
+          data:doc,
+        }),
+        headers:{'content-type':'application/x-www-form-urlencoded'}
+      });
+    },
+
+    /*
+    * 单篇文档放入回收站
+    * */
+    putIntoRecycle:function (doc) {
+      return $http({
+        method:'POST',
+        url:'/admin/manage/document_manage/put_into_recycle',
         data:$.param({
           data:doc,
         }),
@@ -523,58 +533,84 @@ app.controller('categoriesAll', ['$scope', '$http','categoriesAllService', funct
  */
 app.controller('documentAll', ['$scope', '$http', 'documentAllService', function ($scope, $http, documentAllService) {
 
-  documentAllService.get(5, 1).then(function success(res) {
-    $scope.data = res.data.documentByLimitAndPage;
+  documentAllService.postLimitAndPage(5, 1).then(function success(res) {
+    $scope.data = res.data.documentYesDisplayByLimitAndPage;
     $scope.allPage = res.data.allPage;
+    $scope.documentCountNum=res.data.documentCountNum;
   }, function error(res) {
 
   });
 
 
   $scope.limit = '5';
-  $scope.currentPage=1;
+  $scope.currentPage = 1;
   /*
-  * 按条件获取文档数据
-  * */
+   * 按条件获取文档数据
+   * */
   $scope.getPage = function (limit, page) {
     documentAllService.postLimitAndPage(limit, page).then(function success(res) {
-      $scope.data = res.data.documentByLimitAndPage;
+      $scope.data = res.data.documentYesDisplayByLimitAndPage;
       $scope.allPage = res.data.allPage;
-
+      $scope.documentCountNum=res.data.documentCountNum;
     }, function error(res) {
 
     });
-    $scope.currentPage=1;
+    $scope.currentPage = 1;
 
   };
-/*
-* 单击跳转页面
-* */
-  $scope.goToPage=function (limit,page) {
+  /*
+   * 单击跳转页面
+   * */
+  $scope.goToPage = function (limit, page) {
     documentAllService.postLimitAndPage(limit, page).then(function success(res) {
-      $scope.data = res.data.documentByLimitAndPage;
+      $scope.data = res.data.documentYesDisplayByLimitAndPage;
       $scope.allPage = res.data.allPage;
-      $scope.currentPage=page;
+      $scope.documentCountNum=res.data.documentCountNum;
+      $scope.currentPage = page;
     }, function error(res) {
 
     });
   };
-/*
-* 删除单篇文档
-* */
-$scope.removeOneDocument=function (doc) {
-  $scope.oneDocument=doc;
-$('#remove_one_document_modal').modal({
-  keyboard: true
-});
-  documentAllService.removeOneDocument(doc).then(function success(res) {
-    if(res.data.code===1){
+  /*
+   * 删除单篇文档
+   * */
+  $scope.removeOneDocument = function (doc) {
+    $scope.oneDocument = doc;
+    $('#remove_one_document_modal').modal({
+      keyboard: true
+    });
 
+  };
+  /*
+   * 删除单篇文档提交
+   * */
+  $scope.removeOneDocumentCommit = function (doc) {
+
+    $scope.document_display = {
+      name: '1',
+    };
+    console.log($scope.document_display.name);
+    if ($scope.document_display.name == 1) {
+      documentAllService.putIntoRecycle(doc).then(function success(res) {
+        if (res.data.code === 1) {
+          $('#remove_one_document_modal').modal('hide');
+          $scope.getPage();
+        }
+      }, function error(res) {
+
+      });
+    } else {
+      documentAllService.removeOneDocument(doc).then(function success(res) {
+        if (res.data.code === 1) {
+
+        }
+      }, function error(res) {
+
+      });
     }
-  },function error(res) {
 
-  });
-};
+  };
+
 
 }]);
 /**
